+++ src/datasource.rs
@@ -1,13 +1,74 @@
-use crate::{CtxDataType, CtxErrorType};
+use anyhow::{bail, Context, Result};
 use calamine::{open_workbook, CellErrorType, Data, DataType, Reader, Xlsx};
 use csv::{self, Trim};
 use indexmap::IndexMap;
+use serde::Serialize;
 use std::collections::HashMap;
 use std::path::{Path, PathBuf};
+
 pub type DataSourceRows = IndexMap<String, HashMap<String, CtxDataType>>;
-use anyhow::{bail, Context, Result};
+
 pub trait DataSourceReader {
     fn read(&self) -> Result<DataSourceRows>;
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub enum CtxErrorType {
+    /// Division by 0 error
+    Div0,
+    /// Unavailable value error
+    NA,
+    /// Invalid name error
+    Name,
+    /// Null value error
+    Null,
+    /// Number error
+    Num,
+    /// Invalid cell reference error
+    Ref,
+    /// Value error
+    Value,
+    /// Getting data
+    GettingData,
+}
+
+#[derive(Clone, Debug, PartialEq)]
+pub enum CtxDataType {
+    Int(i64),
+    Float(f64),
+    String(String),
+    Bool(bool),
+    DateTime(f64),
+    Error(CtxErrorType),
+    Empty,
+}
+
+impl Serialize for CtxDataType {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        match self {
+            Self::Int(value) => serializer.serialize_i64(*value),
+            Self::Float(value) | Self::DateTime(value) => serializer.serialize_f64(*value),
+            Self::String(value) => serializer.serialize_str(value),
+            Self::Bool(value) => serializer.serialize_bool(*value),
+            Self::Error(value) => {
+                let s = match value {
+                    CtxErrorType::Div0 => "#DIV/0!",
+                    CtxErrorType::NA => "#N/A",
+                    CtxErrorType::Name => "#NAME?",
+                    CtxErrorType::Null => "#NULL!",
+                    CtxErrorType::Num => "#NUM!",
+                    CtxErrorType::Ref => "#REF!",
+                    CtxErrorType::Value => "#VALUE!",
+                    CtxErrorType::GettingData => "#UNKNOWN!",
+                };
+                serializer.serialize_str(s)
+            }
+            Self::Empty => serializer.serialize_unit(),
+        }
+    }
 }
 
 #[derive(Debug, Default)]
