+++ src/lib.rs
@@ -1,65 +1,4 @@
-use serde::Serialize;
-
 pub mod commands;
 pub mod config;
 pub mod datasource;
 pub mod renderer;
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum CtxErrorType {
-    /// Division by 0 error
-    Div0,
-    /// Unavailable value error
-    NA,
-    /// Invalid name error
-    Name,
-    /// Null value error
-    Null,
-    /// Number error
-    Num,
-    /// Invalid cell reference error
-    Ref,
-    /// Value error
-    Value,
-    /// Getting data
-    GettingData,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum CtxDataType {
-    Int(i64),
-    Float(f64),
-    String(String),
-    Bool(bool),
-    DateTime(f64),
-    Error(CtxErrorType),
-    Empty,
-}
-
-impl Serialize for CtxDataType {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match self {
-            Self::Int(value) => serializer.serialize_i64(*value),
-            Self::Float(value) | Self::DateTime(value) => serializer.serialize_f64(*value),
-            Self::String(value) => serializer.serialize_str(value),
-            Self::Bool(value) => serializer.serialize_bool(*value),
-            Self::Error(value) => {
-                let s = match value {
-                    CtxErrorType::Div0 => "#DIV/0!",
-                    CtxErrorType::NA => "#N/A",
-                    CtxErrorType::Name => "#NAME?",
-                    CtxErrorType::Null => "#NULL!",
-                    CtxErrorType::Num => "#NUM!",
-                    CtxErrorType::Ref => "#REF!",
-                    CtxErrorType::Value => "#VALUE!",
-                    CtxErrorType::GettingData => "#UNKNOWN!",
-                };
-                serializer.serialize_str(s)
-            }
-            Self::Empty => serializer.serialize_unit(),
-        }
-    }
-}
